+++ src/dory/input_dg/any_partition/v0/v0_input_dg_reader.cc
@@ -123,6 +123,6 @@ TMsg::TPtr TV0InputDgReader::BuildMsg() {
 
   const uint8_t *value_begin = pos;
   return TryCreateAnyPartitionMsg(ts, topic_begin, topic_end, key_begin,
-      key_sz, value_begin, value_sz, Pool, AnomalyTracker, MsgStateTracker,
-      NoLogDiscard);
+      static_cast<size_t>(key_sz), value_begin, static_cast<size_t>(value_sz),
+      Pool, AnomalyTracker, MsgStateTracker, NoLogDiscard);
 }
